requirements.txt
sqlalchemy
fastapi
pydantic
psycopg2-binary
python-multipart
httpx
pytest
uvicorn
passlib
starlette
python-jose
pytest
